requirements.txt
python-magic==0.4.18
python-magic-bin==0.4.14
pypng==0.0.20